fix(BUG-007): torch_dtype deprecation, log line widths, target-aware probes
──── Three fixes for terminal usability ─────────────────────────────

1. DEPRECATION: torch_dtype= → dtype=  (transformers 4.46+)
   Four call sites updated:
   - train_tokenizer.py:766 (phase_smart_init)
   - finetune.py:1484       (mode_train)
   - finetune.py:2042       (mode_export)
   - analyze_embeddings.py:145 (load_artifacts)

2. LOG WIDTH: narrator messages narrowed to ≤ 80 columns
   Old (117 cols):
     "Step 273: loss stable at ~6.155 (<2% change) — model has
      converged for this lr"
   New (~50 cols):
     "Step 273: loss 6.155 (±2%) — converged"

   Same compression for PROGRESS, WARN-loss-up, grad-clip messages.

3. GRAD WARNING FLOOD: rate-limit grad_norm warnings
   Old: every batch with grad > 1.5 emitted a [WARN]
        → 200+ identical warnings per epoch during cold-start
        → drowns out the useful narration
   New tiered logic:
   - grad > 5.0           → ALWAYS emit (real spike, matters)
   - grad 1.5-5.0, step≤30 → emit every 5 steps (cold-start clipping)
   - grad 1.5-5.0, step>30 → SILENT (clipping is normal & expected)

──── Bonus: target-aware cosine probe table ────────────────────────

The cosine probe output now shows:
- Current value
- Goal target band (per goals.md, e.g. cross-domain 0.0-0.30)
- Status tag (✓ HIT / ⚠ HIGH/LOW / ✗ HIGH/LOW)
- Δ vs previous epoch
- Bar chart with goal band as ▓ and current value as ●

Plus a per-probe insight line explaining WHAT TO DO when off-target:
  - same-tool too high → "variants are clones — differentiate arg shapes"
  - sibling too high   → "indistinguishable — add A3 contrastive lines"
  - cross-domain high  → "DANGER: tools collapsing — add contrastive examples"

Plus a [SUMMARY] line: "X/9 probes IN BAND | Y too high | Z too low"
And direction-of-travel: "M heading toward target, N moving away"
And [ACTION] block listing the 2 biggest gaps to close.

Wired prev_sims through corpus-training epoch end so deltas show:
  same-tool forms     +0.87  0.50-0.80  ⚠ HIGH  -0.04  ↓ heading toward target

PROBE_TARGETS dict at module level is the canonical single source of
truth for goal ranges — sourced from goals.md.

──── Documentation ─────────────────────────────────────────────────

docs/bug-fixes.md BUG-007 entry: deprecation + flood + line width.
Includes before/after table for the message compression and the
tiered rate-limit logic. Lesson: "Verbose telemetry erodes its own
usefulness."

+++ training/finetune.py
@@ -1481,7 +1481,7 @@ def mode_train(epochs: int = 3, lr: float = 2e-4, batch_size: int = 0,
     try:
         model = AutoModelForCausalLM.from_pretrained(
             model_path,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             device_map=device_map,
             trust_remote_code=False,
             low_cpu_mem_usage=True,
@@ -2039,7 +2039,7 @@ def mode_export() -> None:
     try:
         base_model = AutoModelForCausalLM.from_pretrained(
             base_path,
-            torch_dtype=torch.float32,
+            dtype=torch.float32,
             device_map="cpu",
             low_cpu_mem_usage=True,
         )
